Add a simple test that opens an example Salt project and checks if the documents have been loaded
