--- tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/tests/TestProjectManager.java
@@ -1,22 +1,0 @@
-package org.corpus_tools.hexatomic.core.tests;
-
-import org.corpus_tools.hexatomic.core.ProjectManager;
-import org.junit.Before;
-import org.junit.Test;
-
-public class TestProjectManager {
-	
-	private ProjectManager projectManager;
-
-	@Before
-	public void setUp() throws Exception {
-		projectManager = new ProjectManager();
-	}
-
-	@Test(expected = NullPointerException.class)
-	public void testOpenNull() {
-		projectManager.open(null);
-	}
-	
-
-}
